Refactor rain functions to separate module
