+++ tests/test_process.py
@@ -6,16 +6,12 @@ import pandas as pd
 import pytest
 
 from rfactor.process import (
-    _check_path,
     _days_since_start_year,
-    _extract_metadata_from_file_path,
     compute_rainfall_statistics,
     get_rfactor_station_year,
-    load_rain_file,
-    load_rain_file_matlab_legacy,
-    load_rain_folder,
     write_erosivity_data,
 )
+from rfactor.rain import load_rain_file_matlab_legacy, load_rain_folder
 
 
 def test_days_since_last_year_float():
@@ -35,209 +31,6 @@ def test_days_since_last_year_single_year():
     )
     with pytest.raises(Exception):
         _days_since_start_year(ts_series)
-
-
-@pytest.mark.parametrize(
-    "file_name,station,year",
-    [
-        ("UKKEL_2002.txt", "UKKEL", "2002"),
-        ("U_K_K_E_L_1998.txt", "U_K_K_E_L", "1998"),
-        ("U_K_K_E_L_1998.txt", "U_K_K_E_L", "1998"),
-        ("UK-KEL_1998.txt", "UK-KEL", "1998"),
-    ],
-)
-def test_extract_metadata_from_file_path(file_name, station, year):
-    """Split valid file names into station and nam.e"""
-    assert _extract_metadata_from_file_path(Path(file_name)) == (station, year)
-
-
-@pytest.mark.parametrize(
-    "file_name",
-    [
-        "UKKEL_2002---.txt",  # text after year
-        "UKKEL_2002_.txt",  # text after year
-        "UKKEL_202.txt",  # invalid year
-        "UKKEL20022.txt",  # invalid year
-        "2002.txt" "UKKEL.txt",  # no station name  # no year
-    ],
-)
-def test_extract_metadata_from_file_path_invalid(file_name):
-    """Invalid file paths should raise ValueError"""
-    with pytest.raises(ValueError) as excinfo:
-        assert _extract_metadata_from_file_path(Path(file_name))
-    assert "Input file_path_format should" in str(excinfo.value)
-
-
-def test_check_path():
-    """File path checks should provide user with info on pathlib.Path"""
-    assert _check_path(Path("./")) is None
-    with pytest.raises(TypeError) as excinfo:
-        _check_path("invalid_string_input_path")
-    assert "to convert string file_path to valid" in str(excinfo.value)
-    with pytest.raises(TypeError) as excinfo:
-        _check_path(np.array([1, 2]))
-    assert " should be a pathlib.Path" in str(excinfo.value)
-
-
-def test_load_rain_file(rain_data_file):
-    """Valid rainfall data should be parsed to rain DataFrame"""
-    rainfall_data = load_rain_file(rain_data_file, load_rain_file_matlab_legacy)
-    assert isinstance(rainfall_data, pd.DataFrame)
-
-    assert list(rainfall_data.columns) == [
-        "datetime",
-        "station",
-        "rain_mm",
-        "year",
-        "tag",
-    ]
-    assert list(rainfall_data["station"].unique()) == ["station_name"]
-    assert list(rainfall_data.iloc[0].values) == [
-        pd.to_datetime("2021-01-01 00:01:00"),
-        "station_name",
-        1.0,
-        2021,
-        "station_name_2021",
-    ]
-    assert list(rainfall_data.iloc[-1].values) == [
-        pd.to_datetime("2021-12-31 23:59:00"),
-        "station_name",
-        10.00,
-        2021,
-        "station_name_2021",
-    ]
-
-
-class TestCustomLoadRainFile:
-    """Test column name/type checking for custom load functions."""
-
-    def test_dataframe(self, rain_data_file):
-        """Check if it returns a dataframe"""
-
-        def function_no_dataframe(_):
-            return "test"
-
-        with pytest.raises(IOError) as excinfo:
-            load_rain_file(rain_data_file, function_no_dataframe)
-        assert "return pandas.DataFrame" in str(excinfo.value)
-
-    def test_custom_wrong_columns(self, rain_data_file):
-        """Wrong column names return IOError"""
-
-        def function_wrong_columns(_):
-            """Return wrong columns"""
-            return pd.DataFrame(columns=["test"])
-
-        with pytest.raises(IOError) as excinfo:
-            load_rain_file(rain_data_file, function_wrong_columns)
-        assert "must return columns 'datetime', 'station' and 'rain_mm'." in str(
-            excinfo.value
-        )
-
-    def function_wrong_datetime(self, rain_data_file):
-        """Wrong datetime column format (must be datetime64[ns])"""
-
-        def function_wrong_datetime(_):
-            """Return datetime type"""
-            return pd.DataFrame(
-                {"rain_mm": [0.0], "datetime": [0.0], "station": ["str"]}
-            )
-
-        with pytest.raises(IOError) as excinfo:
-            load_rain_file(rain_data_file, function_wrong_datetime)
-        assert "must return datetime64[ns] type" in str(excinfo.value)
-
-    def function_wrong_station(self, rain_data_file):
-        """Wrong station column format (must be str)"""
-
-        def function_wrong_station(_):
-            """Return station type"""
-            return pd.DataFrame(
-                {
-                    "rain_mm": [0.0],
-                    "datetime": pd.to_datetime("01/01/2024 00:00:00"),
-                    "station": [0.0],
-                }
-            )
-
-        with pytest.raises(IOError) as excinfo:
-            load_rain_file(rain_data_file, function_wrong_station)
-        assert "must return object (str) type" in str(excinfo.value)
-
-    def function_wrong_value(self, rain_data_file):
-        """Wrong value column format (must be float)"""
-
-        def function_wrong_value(_):
-            """Return value type"""
-            return pd.DataFrame(
-                {
-                    "rain_mm": [""],
-                    "datetime": pd.to_datetime("01/01/2024 00:00:00"),
-                    "station": [""],
-                }
-            )
-
-        with pytest.raises(IOError) as excinfo:
-            load_rain_file(rain_data_file, function_wrong_value)
-        assert " must return float for column" in str(excinfo.value)
-
-
-def test_load_rain_file_with_folder(rain_data_folder):
-    """When input is a file, should return ValueError to user"""
-
-    with pytest.raises(ValueError) as excinfo:
-        load_rain_file(rain_data_folder, load_rain_file_matlab_legacy)
-    assert "a file instead of a directory" in str(excinfo.value)
-
-
-def test_load_rain_folder(
-    rain_data_folder, data_folder_non_existing, data_folder_empty
-):
-    """Rainfall data should be parsed to rain DataFrame
-    when loading multiple files adding a year and tag column"""
-    rainfall_data = load_rain_folder(rain_data_folder, load_rain_file_matlab_legacy)
-    assert isinstance(rainfall_data, pd.DataFrame)
-    assert list(rainfall_data.columns) == [
-        "datetime",
-        "station",
-        "rain_mm",
-        "year",
-        "tag",
-    ]
-    assert list(rainfall_data["station"].unique()) == ["station_0", "station_1"]
-    assert list(rainfall_data.iloc[0].values) == [
-        pd.to_datetime("2020-01-01 00:01:00"),
-        "station_0",
-        1.0,
-        2020,
-        "station_0_2020",
-    ]
-    assert list(rainfall_data.iloc[-1].values) == [
-        pd.to_datetime("2021-12-31 23:59:00"),
-        "station_1",
-        10.00,
-        2021,
-        "station_1_2021",
-    ]
-    with pytest.raises(FileNotFoundError) as excinfo:
-        load_rain_folder(data_folder_non_existing, load_rain_file_matlab_legacy)
-    assert f"Input folder '{data_folder_non_existing}' does not exists." in str(
-        excinfo.value
-    )
-
-    with pytest.raises(FileNotFoundError) as excinfo:
-        load_rain_folder(data_folder_empty, load_rain_file_matlab_legacy)
-    assert (
-        f"Input folder '{data_folder_empty}' does not contain any 'txt'-files."
-        in str(excinfo.value)
-    )
-
-
-def test_load_rain_folder_with_file(rain_data_file):
-    """When input is a file, should return ValueError to user"""
-    with pytest.raises(ValueError) as excinfo:
-        load_rain_folder(rain_data_file, load_rain_file_matlab_legacy)
-    assert "a directory instead of a file" in str(excinfo.value)
 
 
 def test_write_erosivity(dummy_erosivity, tmpdir):
@@ -314,18 +107,22 @@ def test_rfactor_from_erosivity_subset_not_existing(dummy_erosivity):
     assert "1991" in str(excinfo.value)
 
 
-def test_rainfall_statistics(rain_data_folder):
+def test_rainfall_statistics(rain_data_folder_matlab):
     """"""
-    rainfall_data = load_rain_folder(rain_data_folder, load_rain_file_matlab_legacy)
+    rainfall_data = load_rain_folder(
+        rain_data_folder_matlab, load_rain_file_matlab_legacy
+    )
     rf_stats = compute_rainfall_statistics(rainfall_data)
     assert set(rf_stats.columns) == set(["year", "records", "min", "median", "max"])
     assert isinstance(rf_stats["year"][0], list)
     assert rf_stats["records"].dtype == np.int64
 
 
-def test_rainfall_statistics_with_metadata(rain_data_folder, station_metadata):
+def test_rainfall_statistics_with_metadata(rain_data_folder_matlab, station_metadata):
     """Extend rain stats with metadata"""
-    rainfall_data = load_rain_folder(rain_data_folder, load_rain_file_matlab_legacy)
+    rainfall_data = load_rain_folder(
+        rain_data_folder_matlab, load_rain_file_matlab_legacy
+    )
     rf_stats = compute_rainfall_statistics(rainfall_data, station_metadata)
     assert set(rf_stats.columns) == set(
         ["year", "station", "x", "y", "records", "min", "median", "max"]
